Followed a tutorial
The code that I used is copied from a tutorial, now you can manually play, pause, stop and skip a video as an user. You also see a timer (that resets when a new video starts playing) and you can see which video you are watching out of how many videos there are.